Claude/tech_reviews/requirements.txt
Django==4.2.7
Pillow==10.0.1
django-ckeditor==6.7.0
django-crispy-forms==2.0
django-widget-tweaks
crispy-bootstrap4
requests
